Add first working version of the page
This commit introduces the first version of Broadway dashboard as an
additional page for the Phoenix LiveDashboard.

It works by recording events from the pipelines and displaying a graph
with each stage as a circle.
The circles turn red when they are more busy and green and they are less
busy.

This work is based on Marlus Saraiva's code from his presentation to
ElixirConf 2019. In that talk he presented a graph showing the
work of a Broadway pipeline, which is essentially the same from this
change. See: https://www.youtube.com/watch?v=tPu-P97-cbE