Create Metrics process on demand per pipeline (#2)
This is a simplification of the Metrics GenServer to deal with only one
Pipeline at a time. This way we don't need one server for the entire
pipelines. We can spawn multiple servers and it's OK if they die because
we can spawn again accessing the page.

We also simplified the telemetry module to pattern match the topology
name. This way we don't need branches for detecting the pipeline, and
we also don't need to "parse" the name since now we have it as a
metadata field.