Move counters out of persistent term (#3)
This commit is a refactor that changes the way we keep track of
counters.

We remove the usage of persistent term and instead we are now storing
it in the Metrics process, and also we pass it to our Telemetry
handler. Check the module Counters for details about the implementation.

With this change we swap the strategy of "pulling" the counters data
(in the page process) to a "push" of topology workload from the Metrics
process to our listeners (the page processes). This simplifies a lot
the build of our graphs because we don't need to use RPC for "count"
and "build graph".