Use monotonic time for workload calculations
This is a fix in the workload/idle time calculation that was using the
"system time" from broadway telemetry measurements. The system time does
not work for us because it can "jump" backward in order to fix itself
with the OS and monotonic time.

See https://erlang.org/doc/man/erlang.html#system_time-0 for more
details.

Closes https://github.com/dashbitco/broadway_dashboard/issues/5